Nice_or_Mean.py assignment submission.
Customized Nice_or_Mean.py into a version of Would You Rather allowing for two people to participate. The main player will be asked a series of questions, scoring is based upon whether or not the opponent agrees or disagrees. Defining win and lose as well as giving the opportunity to continue playing or exiting the program.