Inheritance Assignment is submitted.
In class_inheritance.py I create parent class 'Members' and child classes 'Staff' and 'Customers'. Method __innit__() is called to initialize attributes of the class. The child classes have the addition of super().__innit__() to ensure the inherited attributes are properly initialized for instances of child class. This maintains MRO order and will be beneficial int he future if there are multiple inheritances.